Rename flattenjson to flatten
